added "GeneticMemoryNetwork"
the memory network adds a definite amount of neurons to the hidden and output layer of the network. the extra neurons in the output layer take their outputs and apply them to a memory bank to be used by the aditional hiiden layer neurons in later repetitions. The memory network is meant to be used for continuous inputs.